Bump vertical constraint for thumbnails to 843px
In layouts like this, images are constrained horizontally, but not
vertically. We used to retrieve square thumbnails to reduce load on
the IIIF server, but this caused issues with images that are notably
taller than they are wide. Such images would be upscaled to fit into
their horizontal bounds, causing pixelation.

Increasing the vertical bound for the retrieved thumbnail to the
smallest size for copyrighted full-sized images fixes this issue.